trying to get collisions
Added collider to the plan in main scene. Trying to get Destroy or Reset Scene script to work but can't, tried adding collider to TowerBroken02, also tried the script on the terrain collider and can't get anything to happen. Plane does collider with terrain now in the zone with terrain collider. Will work on it more tomorrow